Fix getting end position of Constructs
Getting the end position of the statement that starts the construct just
gives you the line and column that ends the statement, e.g., for DO it
gives you end of "DO'. To get the end position, we have to instead get
the end position of the end statement, stored in a different field of
the tuple.