Fix/cleanup some of the fortran tests
There were some sneaky tab characters that needed to be removed.
Also, there seems to be a bug in flang that causes Sameer's myhi.f
example to have some issues. Basically, you can't have a construct
ended by a continue statement as the last line of a subprogram or
program. Adding a unlabled continue statement between the continue
that terminates the construct and the end statement corresponding to
the procedure or program fixes this.