Fix a (buffer overflow?) bug
There was either a bug in instrumentor.cpp or there is still a bug in
ryml.
When processing the updated config file that includes C and Fortran,
ryml would throw a parse error and show text appended to the last line
of the config file (that's not actually there) in its error output.