Split source location functions into separate file.
The source location routines could be used for purposes other than
instrumentation, so this commit moves them to a separate file and adds a
header so they could be used other than in the instrumentor. Pass the
parser as an argument so that they don't need to be inside the visitor
class.